+++ components/QuestionEdit/index.tsx
@@ -1,20 +1,22 @@
 import React from 'react';
 import Link from 'next/link';
+import sanitizeHtml from 'sanitize-html';
 import styles from './QuestionEdit.module.scss';
-import {Button} from "react-aria-components";
+import { Button } from "react-aria-components";
 
 interface QuestionEditProps {
   id: string;
-  name: string;
+  text: string;
   link: string;
 }
 
 const QuestionEdit: React.FC<QuestionEditProps> = ({
-                                                     id,
-                                                     name,
-                                                     link
-                                                   }) => {
+  id,
+  text,
+  link
+}) => {
 
+  const questionText = sanitizeHtml(text);
   const UpArrowIcon = () => (
     <svg
       width="24"
@@ -57,9 +59,10 @@ const QuestionEdit: React.FC<QuestionEditProps> = ({
         <p className={styles.questionEdit__label} id={`question-label-${id}`}>
           Question
         </p>
-        <p className={styles.questionEdit__name} id={`question-${id}`}>
-          {name}
-        </p>
+        <div className={styles.questionEdit__name} id={`question-${id}`}
+          dangerouslySetInnerHTML={{ __html: questionText }}
+        />
+
       </div>
       <div className={styles.questionEdit__actions} role="group" aria-label="Question actions">
         <Link href={link} className={styles.questionEdit__link} aria-label={`Edit question: ${name}`}>